use u128 instead of i128 if possible

+++ liquidity_pool/src/token.rs
@@ -22,30 +22,30 @@ pub fn create_contract(
         .deploy(token_wasm_hash)
 }
 
-fn get_balance(e: &Env, contract: Address) -> i128 {
+fn get_balance(e: &Env, contract: Address) -> u128 {
     storage::bump_instance(e);
-    Client::new(e, &contract).balance(&e.current_contract_address())
+    Client::new(e, &contract).balance(&e.current_contract_address()) as u128
 }
 
-pub fn get_balance_a(e: &Env) -> i128 {
+pub fn get_balance_a(e: &Env) -> u128 {
     get_balance(e, storage::get_token_a(e))
 }
 
-pub fn get_balance_b(e: &Env) -> i128 {
+pub fn get_balance_b(e: &Env) -> u128 {
     get_balance(e, storage::get_token_b(e))
 }
 
-pub fn get_balance_shares(e: &Env) -> i128 {
+pub fn get_balance_shares(e: &Env) -> u128 {
     get_balance(e, storage::get_token_share(e))
 }
 
-pub fn get_user_balance_shares(e: &Env, user: &Address) -> i128 {
-    Client::new(e, &storage::get_token_share(e)).balance(user)
+pub fn get_user_balance_shares(e: &Env, user: &Address) -> u128 {
+    Client::new(e, &storage::get_token_share(e)).balance(user) as u128
 }
 
-pub fn get_total_shares(e: &Env) -> i128 {
+pub fn get_total_shares(e: &Env) -> u128 {
     let share_token = storage::get_token_share(e);
-    Client::new(e, &share_token).total_balance()
+    Client::new(e, &share_token).total_balance() as u128
 }
 
 pub fn burn_shares(e: &Env, amount: i128) {
@@ -62,10 +62,10 @@ fn transfer(e: &Env, token: Address, to: Address, amount: i128) {
     Client::new(e, &token).transfer(&e.current_contract_address(), &to, &amount);
 }
 
-pub fn transfer_a(e: &Env, to: Address, amount: i128) {
-    transfer(e, storage::get_token_a(e), to, amount);
+pub fn transfer_a(e: &Env, to: Address, amount: u128) {
+    transfer(e, storage::get_token_a(e), to, amount as i128);
 }
 
-pub fn transfer_b(e: &Env, to: Address, amount: i128) {
-    transfer(e, storage::get_token_b(e), to, amount);
+pub fn transfer_b(e: &Env, to: Address, amount: u128) {
+    transfer(e, storage::get_token_b(e), to, amount as i128);
 }
